requirements.txt
# Core NLP and Machine Learning
spacy>=3.7.0                    # Advanced NLP and Named Entity Recognition
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# PDF and Document Processing
pymupdf>=1.23.0                 # PDF text extraction (PyMuPDF)
python-docx>=1.1.0              # DOCX file processing

# Web Application Framework
streamlit>=1.28.0               # Interactive web UI

# Data Processing and Analysis
pandas>=2.1.0                   # Data manipulation and analysis
numpy>=1.25.0                   # Numerical computing

# Export Capabilities
openpyxl>=3.1.0                 # Excel file creation
xlsxwriter>=3.1.0               # Advanced Excel formatting
